function/rehydrate function/requirements.txt
azure-functions
azure-identity
azure-keyvault-secrets
azure-cosmos
azure-storage-blob
